fix(desktop): rename compose.ts/messageList.ts to avoid case-only filename collisions
Compose.tsx + compose.ts (and MessageList.tsx + messageList.ts) differ only in
case. On case-insensitive filesystems (macOS) TS module resolution picks the .ts
over the .tsx for './Compose', so the compose modal + ComposeInit type became
unreachable and 'tsc && vite build' (via 'wails dev') failed with a cascade of
'differs only in casing' / 'no default export' / implicit-any errors. Linux CI
is case-sensitive, so it passed there.

Rename the pure-logic modules to non-colliding names:
- compose.ts -> composeBuilders.ts (+ its test)
- messageList.ts -> messageListModel.ts (+ its test)
and update importers. The PascalCase components (Compose.tsx, MessageList.tsx)
now resolve unambiguously. tsc + build + 61 unit + 39 e2e green.
